api: rename engines.py → conversions.py; update roadmap checkboxes (#108)
Rename:
- src/ad_hoc_diffractometer/engines.py → conversions.py
- tests/test_engines.py → test_conversions.py

Update all imports, cross-references, and documentation:
- __init__.py: from .engines → from .conversions
- api.rst: graph node 'engines' → 'conversions'; card text; mod reference
- howto/wavelength.md: mod cross-reference
- conversions.py: module docstring filename

roadmap.md checkboxes updated for work completed in this session:
- §3.8: all energy/wavenumber/named-lines items checked (done)
- §3.8e: notebook rendering (myst-nb) and User Guide items checked
- §3.8g: factory basis argument items checked (done)
- §3.9: all neutron source items checked (done)
- §2.3: setor1/setor2 updated to setor0/setor1

1203 tests, 100% coverage. Zero doc build warnings.

Contributed by: OpenCode (argo/claudesonnet46)
