+++ src/ad_hoc_diffractometer/__init__.py
@@ -14,21 +14,21 @@ from .axes import parse_axis
 from .constants import XHAT
 from .constants import YHAT
 from .constants import ZHAT
+from .conversions import Q_mag_to_two_theta
+from .conversions import Q_to_d
+from .conversions import Q_to_hkl
+from .conversions import d_to_Q_mag
+from .conversions import d_to_two_theta
+from .conversions import hkl_to_d
+from .conversions import hkl_to_Q
+from .conversions import hkl_to_two_theta
+from .conversions import two_theta_to_d
+from .conversions import two_theta_to_Q_mag
 from .display import allclose
 from .display import fmt
 from .display import get_precision
 from .display import precision_atol
 from .display import set_precision
-from .engines import Q_mag_to_two_theta
-from .engines import Q_to_d
-from .engines import Q_to_hkl
-from .engines import d_to_Q_mag
-from .engines import d_to_two_theta
-from .engines import hkl_to_d
-from .engines import hkl_to_Q
-from .engines import hkl_to_two_theta
-from .engines import two_theta_to_d
-from .engines import two_theta_to_Q_mag
 from .factories import BASIS_BL
 from .factories import BASIS_YOU
 from .factories import GEOMETRY_ENTRY_POINT_GROUP
